pyproject.toml
[project]
name = "rag-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv==2.2.0",
    "chromadb==1.0.20",
    "docx2txt==0.9",
    "faiss-cpu==1.12.0",
    "hf-xet==1.1.9",
    "ipykernel>=6.30.1",
    "jq==1.10.0",
    "langchain==0.3.27",
    "langchain-astradb==0.6.1",
    "langchain-community==0.3.28",
    "langchain-experimental==0.3.4",
    "langchain-groq==0.3.7",
    "langchain-huggingface==0.3.1",
    "langchain-openai==0.3.32",
    "langchain-pinecone==0.2.11",
    "langchain-tavily==0.2.11",
    "langgraph==0.6.7",
    "langgraph-cli[inmem]>=0.4.2",
    "matplotlib==3.10.6",
    "msoffcrypto-tool==5.4.2",
    "openpyxl==3.1.5",
    "pandas==2.3.2",
    "pymupdf==1.26.4",
    "pypdf==6.0.0",
    "python-docx==1.2.0",
    "python-dotenv==1.1.1",
    "rank-bm25==0.2.2",
    "sentence-transformers==5.1.0",
    "tavily-python==0.7.12",
    "tiktoken==0.11.0",
    "unstructured==0.18.14",
    "wikipedia==1.4.0",
]
